move the `get_z_sources_sinks` method from the `SourceSinkConverter` to the `PolyFile` class
